+++ src/main/java/f18a14c09s/integration/alexa/music/catalog/data/AbstractCatalog.java
@@ -2,6 +2,7 @@ package f18a14c09s.integration.alexa.music.catalog.data;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import f18a14c09s.integration.alexa.data.Locale;
+import f18a14c09s.integration.alexa.music.entities.BaseEntity;
 import lombok.Getter;
 import lombok.Setter;
