+++ src/main/java/f18a14c09s/integration/alexa/music/catalog/data/MusicGroupCatalog.java
@@ -1,6 +1,7 @@
 package f18a14c09s.integration.alexa.music.catalog.data;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import f18a14c09s.integration.alexa.music.entities.Artist;
 import lombok.Getter;
 import lombok.Setter;
