+++ src/main/java/f18a14c09s/integration/alexa/music/catalog/data/MusicPlaylistCatalog.java
@@ -1,6 +1,7 @@
 package f18a14c09s.integration.alexa.music.catalog.data;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import f18a14c09s.integration.alexa.music.entities.Playlist;
 import lombok.Getter;
 import lombok.Setter;
