+++ src/main/java/f18a14c09s/integration/alexa/music/entities/AlbumReference.java
@@ -1,4 +1,4 @@
-package f18a14c09s.integration.alexa.music.catalog.data;
+package f18a14c09s.integration.alexa.music.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
@@ -9,8 +9,9 @@ import java.util.*;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ArtistReference {
+public class AlbumReference {
     private String id;
     private ArrayList<EntityName> names;
     private ArrayList<AlternateNames> alternateNames;
+    private String releaseType;
 }
