+++ src/main/java/f18a14c09s/integration/alexa/music/entities/ArtistReference.java
@@ -1,4 +1,4 @@
-package f18a14c09s.integration.alexa.music.catalog.data;
+package f18a14c09s.integration.alexa.music.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
@@ -9,9 +9,8 @@ import java.util.*;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class AlbumReference {
+public class ArtistReference {
     private String id;
     private ArrayList<EntityName> names;
     private ArrayList<AlternateNames> alternateNames;
-    private String releaseType;
 }
