+++ src/main/java/f18a14c09s/integration/alexa/music/entities/CatalogTypeName.java
@@ -1,4 +1,4 @@
-package f18a14c09s.integration.alexa.music.catalog.data;
+package f18a14c09s.integration.alexa.music.entities;
 
 public class CatalogTypeName {
     public static final String AMAZON_BROADCAST_CHANNEL = "AMAZON.BroadcastChannel";
