+++ src/main/java/f18a14c09s/integration/alexa/music/entities/Playlist.java
@@ -1,4 +1,4 @@
-package f18a14c09s.integration.alexa.music.catalog.data;
+package f18a14c09s.integration.alexa.music.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import java.util.*;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Genre extends BaseEntity {
+public class Playlist extends BaseEntity {
     private ArrayList<EntityName> names;
     private Popularity popularity;
     private ArrayList<AlternateNames> alternateNames;
